refactor: cartItem mock 데이터 폴더 구조 수정
