+++ src/recoil/cartItems/cartItems.test.tsx
@@ -5,7 +5,7 @@ import { RecoilRoot } from 'recoil';
 
 import { cartItemsState } from './atoms';
 
-import { TOTAL_PRICE_OVER_100000_DATA, TOTAL_PRICE_UNDER_100000_DATA } from '@/constants/mock';
+import { TOTAL_PRICE_OVER_100000_DATA, TOTAL_PRICE_UNDER_100000_DATA } from '@/mocks/cartItems';
 import CartProducts from '@components/Cart/CartProducts';
 import Loading from '@components/common/Loading';
